# Conflicts:
#	docs/qdownload.md

+++ iqshell/storage/object/download/downloader.go
@@ -15,9 +15,10 @@ import (
 	"net/http"
 	"os"
 	"strings"
+	"sync"
 )
 
-type ApiInfo struct {
+type DownloadActionInfo struct {
 	Bucket                 string            `json:"bucket"`               // 文件所在 bucket 【必填】
 	Key                    string            `json:"key"`                  // 文件被保存的 key 【必填】
 	IsPublic               bool              `json:"-"`                    // 是否使用公有链接 【必填】
@@ -29,6 +30,8 @@ type ApiInfo struct {
 	ServerFilePutTime      int64             `json:"server_file_put_time"` // 文件修改时间 【选填】
 	ServerFileSize         int64             `json:"server_file_size"`     // 文件大小，有值则会检测文件大小 【选填】
 	ServerFileHash         string            `json:"server_file_hash"`     // 文件 hash，有值则会检测 hash 【选填】
+	DownloadFileSize       int64             `json:"download_file_size"`   // 下载的文件大小，下载整个文件时，等于 ServerFileSize；切片下载则为切片大小；有值则会检测文件大小【选填】
+	CheckSize              bool              `json:"-"`                    // 是否检测文件大小 【选填】
 	CheckHash              bool              `json:"-"`                    // 是否检测文件 hash 【选填】
 	FromBytes              int64             `json:"-"`                    // 下载开始的位置，内部会缓存 【内部使用】
 	ToBytes                int64             `json:"-"`                    // 下载的终止位置【内部使用】
@@ -41,25 +44,25 @@ type ApiInfo struct {
 	Progress               progress.Progress `json:"-"`                    // 下载进度回调【选填】
 }
 
-func (i *ApiInfo) WorkId() string {
+func (i *DownloadActionInfo) WorkId() string {
 	return fmt.Sprintf("%s:%s:%s", i.Bucket, i.Key, i.ToFile)
 }
 
-type ApiResult struct {
+type DownloadActionResult struct {
 	FileModifyTime int64  `json:"file_modify_time"` // 下载后文件修改时间
 	FileAbsPath    string `json:"file_abs_path"`    // 文件被保存的绝对路径
 	IsUpdate       bool   `json:"is_update"`        // 是否为接续下载
 	IsExist        bool   `json:"is_exist"`         // 是否为已存在
 }
 
-var _ flow.Result = (*ApiResult)(nil)
+var _ flow.Result = (*DownloadActionResult)(nil)
 
-func (a *ApiResult) IsValid() bool {
+func (a *DownloadActionResult) IsValid() bool {
 	return len(a.FileAbsPath) > 0 && a.FileModifyTime > 0
 }
 
 // Download 下载一个文件，从 Url 下载保存至 ToFile
-func Download(info *ApiInfo) (res *ApiResult, err *data.CodeError) {
+func Download(info *DownloadActionInfo) (res *DownloadActionResult, err *data.CodeError) {
 	if len(info.ToFile) == 0 {
 		err = data.NewEmptyError().AppendDesc("the filename saved after downloading is empty")
 		return
@@ -70,7 +73,7 @@ func Download(info *ApiInfo) (res *ApiResult, err *data.CodeError) {
 		return
 	}
 
-	res = &ApiResult{
+	res = &DownloadActionResult{
 		FileAbsPath: f.toAbsFile,
 	}
 
@@ -89,9 +92,11 @@ func Download(info *ApiInfo) (res *ApiResult, err *data.CodeError) {
 	}
 
 	// 文件存在则检查文件状态
-	checkMode := object.MatchCheckModeFileSize
+	checkMode := -1
 	if info.CheckHash {
 		checkMode = object.MatchCheckModeFileHash
+	} else if info.CheckSize {
+		checkMode = object.MatchCheckModeFileSize
 	}
 
 	// 读不到 status 按不存在该文件处理
@@ -99,30 +104,36 @@ func Download(info *ApiInfo) (res *ApiResult, err *data.CodeError) {
 	tempFileStatus, _ := os.Stat(f.tempFile)
 	if fileStatus != nil {
 		// 文件已下载，检测文件内容
-		checkResult, mErr := object.Match(object.MatchApiInfo{
-			Bucket:         info.Bucket,
-			Key:            info.Key,
-			LocalFile:      f.toAbsFile,
-			CheckMode:      checkMode,
-			ServerFileHash: info.ServerFileHash,
-			ServerFileSize: info.ServerFileSize,
-		})
-		if mErr != nil {
-			f.fromBytes = 0
-			log.DebugF("check error before download:%v", mErr)
-		}
-		if checkResult != nil {
-			res.IsExist = checkResult.Exist
-		}
-
-		if mErr == nil && checkResult.Match {
-			// 文件已下载，并在文件匹配，不再下载
-			if fileModifyTime, fErr := utils.LocalFileModify(f.toAbsFile); fErr != nil {
-				log.WarningF("Get file ModifyTime error:%v", fErr)
-			} else {
-				res.FileModifyTime = fileModifyTime
+		if checkMode < 0 {
+			// 文件已存在，无论文件是什么均认为是预期
+			res.IsExist = true
+			return res, nil
+		} else {
+			checkResult, mErr := object.Match(object.MatchApiInfo{
+				Bucket:         info.Bucket,
+				Key:            info.Key,
+				LocalFile:      f.toAbsFile,
+				CheckMode:      checkMode,
+				ServerFileHash: info.ServerFileHash,
+				ServerFileSize: info.DownloadFileSize,
+			})
+			if mErr != nil {
+				f.fromBytes = 0
+				log.DebugF("check error before download:%v", mErr)
 			}
-			return
+			if checkResult != nil {
+				res.IsExist = checkResult.Exist
+			}
+
+			if mErr == nil && checkResult.Match {
+				// 文件已下载，并在文件匹配，不再下载
+				if fileModifyTime, fErr := utils.LocalFileModify(f.toAbsFile); fErr != nil {
+					log.WarningF("Get file ModifyTime error:%v", fErr)
+				} else {
+					res.FileModifyTime = fileModifyTime
+				}
+				return
+			}
 		}
 	} else if tempFileStatus != nil && tempFileStatus.Size() > 0 {
 		// 文件已下载了一部分，需要继续下载
@@ -159,26 +170,28 @@ func Download(info *ApiInfo) (res *ApiResult, err *data.CodeError) {
 	}
 
 	// 检查下载后的数据是否符合预期
-	checkResult, mErr := object.Match(object.MatchApiInfo{
-		Bucket:         info.Bucket,
-		Key:            info.Key,
-		LocalFile:      f.toAbsFile,
-		CheckMode:      checkMode,
-		ServerFileHash: info.ServerFileHash,
-		ServerFileSize: info.ServerFileSize,
-	})
-	if mErr != nil || (checkResult != nil && !checkResult.Match) {
-		log.DebugF("after download, remove download file because file doesn't match")
-		if rErr := os.Remove(f.toAbsFile); rErr != nil {
-			log.ErrorF("after download, remove download file error:%s", rErr)
+	if checkMode >= 0 {
+		checkResult, mErr := object.Match(object.MatchApiInfo{
+			Bucket:         info.Bucket,
+			Key:            info.Key,
+			LocalFile:      f.toAbsFile,
+			CheckMode:      checkMode,
+			ServerFileHash: info.ServerFileHash,
+			ServerFileSize: info.DownloadFileSize,
+		})
+		if mErr != nil || (checkResult != nil && !checkResult.Match) {
+			log.DebugF("after download, remove download file because file doesn't match")
+			if rErr := os.Remove(f.toAbsFile); rErr != nil {
+				log.ErrorF("after download, remove download file error:%s", rErr)
+			}
+			return res, data.NewEmptyError().AppendDesc("check error after download").AppendError(mErr)
 		}
-		return res, data.NewEmptyError().AppendDesc("check error after download").AppendError(mErr)
 	}
 
 	return res, nil
 }
 
-func download(fInfo *fileInfo, info *ApiInfo) (err *data.CodeError) {
+func download(fInfo *fileInfo, info *DownloadActionInfo) (err *data.CodeError) {
 	defer func() {
 		if info.RemoveTempWhileError && err != nil {
 			e := os.Remove(fInfo.tempFile)
@@ -191,66 +204,102 @@ func download(fInfo *fileInfo, info *ApiInfo) (err *data.CodeError) {
 	}()
 
 	info.FromBytes = fInfo.fromBytes
-	err = downloadFile(fInfo, info)
+	err = downloadTempFile(fInfo, info)
 	if err != nil {
 		return err
 	}
 
-	err = renameTempFile(fInfo, info)
+	err = renameTempFile(fInfo)
 	return err
 }
 
-func downloadFile(fInfo *fileInfo, info *ApiInfo) *data.CodeError {
-	dl, err := createDownloader(info)
-	if err != nil {
-		return data.NewEmptyError().AppendDesc("Download create downloader error:" + err.Error())
-	}
-
-	var response *http.Response
+func downloadTempFile(fInfo *fileInfo, info *DownloadActionInfo) (err *data.CodeError) {
+	useHttps := workspace.GetConfig().IsUseHttps()
 	for times := 0; times < 6; times++ {
+		dl, cErr := createDownloader(info)
+		if cErr != nil {
+			return data.NewEmptyError().AppendDesc("Download create downloader error:" + cErr.Error())
+		}
 		log.DebugF("Download[%d] [%s:%s] => %s", times, info.Bucket, info.Key, info.ToFile)
 
-		if available, aErr := info.HostProvider.Available(); !available {
-			err = data.NewEmptyError().AppendDescF("no available host:%+v", aErr)
-			log.DebugF("Stop download [%s:%s] => %s, because no available host", info.Bucket, info.Key, info.ToFile)
+		h, pErr := info.HostProvider.Provide()
+		if h == nil || pErr != nil {
+			if err == nil {
+				err = data.NewEmptyError().AppendDescF("no available host:%+v", pErr)
+			}
+			log.DebugF("Stop download [%s:%s] => %s, because %v", info.Bucket, info.Key, info.ToFile, err)
 			break
 		}
 
-		err = nil
-		response = nil
-		response, err = dl.Download(info)
-		if err == nil && response != nil && response.StatusCode/100 == 2 {
-			err = nil
+		hostString := h.GetServer()
+		urlString, cErr := createDownloadUrlWithHost(h, info, useHttps)
+		if len(hostString) == 0 || cErr != nil {
+			err = data.NewEmptyError().AppendDescF("create download url error:%+v", cErr)
+			log.DebugF("Stop download [%s:%s] => %s, because %+v", info.Bucket, info.Key, info.ToFile, err)
 			break
 		}
 
-		if response != nil {
-			log.DebugF("Download[%d] [%s:%s] => %s, response:%+v", times, info.Bucket, info.Key, info.ToFile, response)
-			if (response.StatusCode > 399 && response.StatusCode < 500) ||
-				response.StatusCode == 612 || response.StatusCode == 631 {
-				log.DebugF("Stop download [%s:%s] => %s, because [%s]", info.Bucket, info.Key, info.ToFile, response.Status)
-				break
-			}
+		err = downloadTempFileWithDownloader(dl, fInfo, &DownloadApiInfo{
+			Url:            urlString,
+			Host:           hostString,
+			Referer:        info.Referer,
+			RangeFromBytes: fInfo.fromBytes,
+			RangeToBytes:   0,
+			CheckSize:      info.CheckSize,
+			FileSize:       info.ServerFileSize,
+			CheckHash:      info.CheckHash,
+			FileHash:       info.ServerFileHash,
+			Progress:       info.Progress,
+		})
+		if err == nil {
+			break
 		}
 
-		if err != nil {
-			log.DebugF("Download[%d] [%s:%s] => %s, error:%+v", times, info.Bucket, info.Key, info.ToFile, err)
-			if err.Code == 0 || err.IsCancel() {
-				log.DebugF("Stop download [%s:%s] => %s, because [%+v]", info.Bucket, info.Key, info.ToFile, err)
-				break
-			}
+		log.DebugF("Download[%d] [%s:%s] => %s, err:%+v", times, info.Bucket, info.Key, info.ToFile, err)
+		if (err.Code > 399 && err.Code < 500) ||
+			err.Code == 612 || err.Code == 631 {
+			log.DebugF("Stop download [%s:%s] => %s, because [%+v]", info.Bucket, info.Key, info.ToFile, err)
+			break
+		}
+
+		info.HostProvider.Freeze(h)
+		log.DebugF("download freeze host:%s", hostString)
+	}
+	return err
+}
+
+func downloadTempFileWithDownloader(dl downloader, fInfo *fileInfo, info *DownloadApiInfo) *data.CodeError {
+	// 下载之前先验证文件大小，两点考虑：
+	// 1. 文件可能被更换，非预期文件
+	// 2. 文件开启瘦身等，预期文件
+	// 上面两点无法区分，但必须能让用户可以下载预期文件
+	// 不检测文件信息时，使用下载的文件信息作为标准，可以保证下载成功
+	// 此方案有个问题：如果获取文件信息之后，下载之前文件改变了，下载仍会失败（此情景概率极低，且用户重新下载即可）。
+	if file, err := utils.GetNetworkFileInfo(info.Url); err != nil {
+		return err
+	} else if info.CheckHash && info.FileHash != file.Hash {
+		return data.NewEmptyError().AppendDescF("file(%s) hash doesn't match, %s but except:%s", info.Url, file.Hash, info.FileHash)
+	} else if info.CheckSize && info.FileSize != file.Size {
+		return data.NewEmptyError().AppendDescF("file(%s) size doesn't match, %d but except:%d", info.Url, file.Size, info.FileSize)
+	} else {
+		info.FileSize = file.Size
+		info.FileHash = file.Hash
+	}
+
+	// 检查 fromBytes 和 fileSize，fromBytes 不能 > fileSize
+	if info.RangeFromBytes > 0 {
+		if info.RangeFromBytes > info.FileSize || info.RangeFromBytes > info.RangeToBytes {
+			errorDesc := "download, check fromBytes error: fromBytes bigger than file size, should remove temp file and retry."
+			log.Error(errorDesc)
+			_ = fInfo.cleanTempFile()
+			return data.NewEmptyError().AppendDesc(errorDesc)
+		} else if info.RangeFromBytes == info.FileSize || info.RangeFromBytes == info.RangeToBytes {
+			// 已经完全下载，只不过未修改名称
+			return nil
 		}
 	}
 
-	if response != nil && response.Body != nil {
-		if info.Progress != nil {
-			info.Progress.SetFileSize(response.ContentLength + info.FromBytes)
-			info.Progress.SendSize(info.FromBytes)
-			info.Progress.Start()
-		}
-		defer response.Body.Close()
-	}
-
+	response, err := dl.Download(info)
 	if err != nil {
 		return data.NewEmptyError().AppendDesc(" Download error:" + err.Error())
 	}
@@ -258,10 +307,19 @@ func downloadFile(fInfo *fileInfo, info *ApiInfo) *data.CodeError {
 		return data.NewEmptyError().AppendDesc(" Download error: response empty")
 	}
 	if response.StatusCode/100 != 2 {
-		return data.NewEmptyError().AppendDescF(" Download error: %v", response)
+		return data.NewError(response.StatusCode, "").AppendDescF(" Download error: %v", response)
 	}
 	if response.Body == nil {
 		return data.NewEmptyError().AppendDesc(" Download error: response body empty")
+	}
+
+	if response != nil && response.Body != nil {
+		if info.Progress != nil {
+			info.Progress.SetFileSize(info.FileSize)
+			info.Progress.SendSize(info.RangeFromBytes)
+			info.Progress.Start()
+		}
+		defer response.Body.Close()
 	}
 
 	var fErr error
@@ -269,7 +327,7 @@ func downloadFile(fInfo *fileInfo, info *ApiInfo) *data.CodeError {
 	isExist, _ := utils.ExistFile(fInfo.tempFile)
 	if isExist {
 		tempFileHandle, fErr = os.OpenFile(fInfo.tempFile, os.O_APPEND|os.O_WRONLY, 0655)
-		log.DebugF("download [%s:%s] => %s from:%d", info.Bucket, info.Key, info.ToFile, info.FromBytes)
+		log.DebugF("download %s => %s from:%d", info.Url, fInfo.toFile, info.RangeFromBytes)
 	} else {
 		tempFileHandle, fErr = os.Create(fInfo.tempFile)
 	}
@@ -293,7 +351,7 @@ func downloadFile(fInfo *fileInfo, info *ApiInfo) *data.CodeError {
 	return nil
 }
 
-func renameTempFile(fInfo *fileInfo, info *ApiInfo) *data.CodeError {
+func renameTempFile(fInfo *fileInfo) *data.CodeError {
 	err := os.Rename(fInfo.tempFile, fInfo.toAbsFile)
 	if err != nil {
 		return data.NewEmptyError().AppendDescF(" Rename temp file to final file error:%v", err.Error())
@@ -302,12 +360,10 @@ func renameTempFile(fInfo *fileInfo, info *ApiInfo) *data.CodeError {
 }
 
 type downloader interface {
-	Download(info *ApiInfo) (response *http.Response, err *data.CodeError)
+	Download(info *DownloadApiInfo) (response *http.Response, err *data.CodeError)
 }
 
-func createDownloader(info *ApiInfo) (downloader, *data.CodeError) {
-	userHttps := workspace.GetConfig().IsUseHttps()
-
+func createDownloader(info *DownloadActionInfo) (downloader, *data.CodeError) {
 	// 使用切片，并发至少为 2，至少要能切两片，达到切片阈值
 	if info.EnableSlice &&
 		info.SliceConcurrentCount > 1 &&
@@ -315,27 +371,20 @@ func createDownloader(info *ApiInfo) (downloader, *data.CodeError) {
 		info.ServerFileSize > info.SliceFileSizeThreshold {
 		return &sliceDownloader{
 			SliceSize:              info.SliceSize,
+			FileHash:               info.ServerFileHash,
+			FileEncoding:           info.FileEncoding,
+			ToFile:                 info.ToFile,
 			slicesDir:              "",
-			concurrentCount:        info.SliceConcurrentCount,
+			ConcurrentCount:        info.SliceConcurrentCount,
 			totalSliceCount:        0,
 			slices:                 nil,
 			downloadError:          nil,
 			currentReadSliceIndex:  0,
 			currentReadSliceOffset: 0,
+			locker:                 sync.Mutex{},
 		}, nil
 	} else {
-		if info.UseGetFileApi {
-			mac, err := workspace.GetMac()
-			if err != nil {
-				return nil, data.NewEmptyError().AppendDescF("download get mac error:%v", mac)
-			}
-			return &getFileApiDownloader{
-				useHttps: userHttps,
-				mac:      mac,
-			}, nil
-		} else {
-			return &getDownloader{useHttps: userHttps}, nil
-		}
+		return &downloaderFile{}, nil
 	}
 }
